worker/requirements.txt
boto3==1.34.130
botocore==1.34.130
python-dotenv==1.0.1
Django==5.0.7        # para usar o ORM do backend
djangorestframework==3.15.2
psycopg2-binary==2.9.9
paddleocr==3.2.0
paddlepaddle==3.2.0
pillow
numpy